initial commit
Initialize the server. The server uses Express and is split into two routing modules, one for the MVP application and one for the V2 application. Each module has a route to retrieve the reviews and a route to add a new review. There is no database integration in this version of the server.