database setup and integration
I've opted to use MySQL for my database. I've added a .env configuration file to store the MySQL credentials and other database information and I've added a file to store all functions that interface with MySQL directly, which includes functions for getting the reviews, adding a review, and calculating the average rating for a given review table. The MVP router used by express has been updated to use the database integration and all static data has been removed. Added a "/reset" route to the MVP router to allow me to reset the database to its initial state after testing. This route is not exposed to the MVP web page.