Add product table and error handling
After re-reading the challenge, I noticed that I was supposed to include a product table, so I added that table and integrated it with the reviews tables. I also added error handling to the server so that any errors are handled by the server and reported back to the user rather than just letting the server crash. Also slightly changed the MVP API so it is more obvious what each endpoint does and added an endpoint to retrieve the reviews for only one product.